Result set ordering in getQueryResult
When string-keyed arrays are returned from PHP to Lua, they are turned into string-keyed tables. However, the string-keyed table in Lua is not guaranteed to keep its elements in the same order as they were in the PHP array. Thus, any sort applied to the result set is lost. By changing the result set to an integer-keyed array, the ordering can be guaranteed. The keys begin at 1 in order to conform to Lua's expectations regarding integer keys in a table.